requirements.txt
langchain
langchain-groq
langchain-chroma
langchain-community
streamlit
python-dotenv
pysqlite3-binary